wip
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3478 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/lights_b040laserbox/qfextensionb040laserbox.cpp
@@ -94,7 +94,7 @@ void QFExtensionB040LaserBox::deinit() {
     QSettings inifile(services->getGlobalConfigFileDirectory()+QString("/%1").arg(GLOBAL_CONFIGFILE), QSettings::IniFormat);
     if (inifile.isWritable()) {
         inifile.setValue("lightsource_count", getLightSourceCount());
-        for (unsigned int i=0; i<sources.size(); i++) {
+        for (unsigned int i=0; (int64_t)i<sources.size(); i++) {
             int p=sources[i].port;
             if (ports.getCOMPort(p)) {
                 ports.storeCOMPort(p, inifile, "lightsource"+QString::number(i+1)+"/");
